Fix int identity overflow and improve SmartUploadImportService reliability
Change telecom table Id columns (Safaricom, Airtel, PSTN, PrivateWire)
from int to bigint to prevent identity overflow. Add EF migration with
raw SQL ALTER statements. Improve import service with: retry duplicate
prevention via ImportJobId cleanup, error threshold abort (>1000 errors),
CancellationToken threading to SqlBulkCopy, connection reuse across
batches, header-not-found throws, nullable DataTable columns, unified
BulkInsertAsync method, and ImportConstants for magic strings/numbers.

+++ Pages/Admin/CallLogs.cshtml.cs
@@ -3676,8 +3676,8 @@ namespace TAB.Web.Pages.Admin
             }
 
             var deletedCount = 0;
-            var failedRecords = new List<int>();
-            var stagedRecords = new List<int>();
+            var failedRecords = new List<long>();
+            var stagedRecords = new List<long>();
 
             try
             {
@@ -3808,7 +3808,7 @@ namespace TAB.Web.Pages.Admin
         public class BulkDeleteRequest
         {
             public string Type { get; set; } = string.Empty;
-            public List<int> Ids { get; set; } = new();
+            public List<long> Ids { get; set; } = new();
         }
 
         public async Task<IActionResult> OnPostDeleteImportBatchAsync(int importId)
